+++ CHANGELOG.md
@@ -101,6 +101,7 @@
 81. Fixed Issue with Exceptions not including StackTrace.
 82. Fixed Issue with certain Mono Games not having the FileVersionInfo class.
 83. Added Value Change Callbacks to MelonPreferences Entries.
+84. Cleaned up MelonPreferences Entry Exception Handling.
 
 ---